Password reset, new entites & more

+++ src/main/java/ua/edu/ukma/mandarin/scheduler/domain/dto/SubjectDTO.java
@@ -1,16 +1,15 @@
 package ua.edu.ukma.mandarin.scheduler.domain.dto;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-
-import java.util.List;
 
 @AllArgsConstructor
 @Getter
 public class SubjectDTO {
-    private Long id;
-    private String name;
-    private int authorId;
-    //Subject should be added with groups usually
-    private List<GroupDTO> groupDTOList;
+  private Long id;
+  private String name;
+  private Long authorId;
+  // Subject should be added with groups usually
+  private List<GroupDTO> groupDTOList;
 }
